refactor: 💡 move predict.py to correct folder
